Use a logError() that actually works
I just realized the logError() that was there before is actually not going to ever DM anyone. Now it tries its best to get a hold of a Discord client that it can DM from. While I was at it stopped hardcoding my own user ID and made it DM from the ADMINS environment variable instead.